Налажена структура файлов и импортов
