+++ js/utils/utils.js
@@ -1,3 +1,3 @@
-calculateDistance(x1, y1, x2, y2) {
+export function calculateDistance(x1, y1, x2, y2) {
     return Math.sqrt(Math.pow(x1 - x2, 2) + Math.pow(y1 - y2, 2))
 }
